Made another HandlerFactory for FirstRouter.
